Rename monthly summary CSV module
